Chain: added ChainFactory to create chain/step/calculation objects
