+++ webapp/views/App/views/Analysis/Chain/store/actions/calculation/useCreate.js
@@ -1,6 +1,7 @@
 import { useCallback } from 'react'
 import * as A from '@core/arena'
 
+import * as ChainFactory from '@common/analysis/chainFactory'
 import * as Chain from '@common/analysis/processingChain'
 import * as Step from '@common/analysis/processingStep'
 
@@ -9,7 +10,7 @@ import { State } from '../../state'
 export const useCreate = ({ setState }) => {
   return useCallback(({ state }) => {
     const step = State.getStepEdit(state)
-    const calculation = Chain.newProcessingStepCalculation(step)
+    const calculation = ChainFactory.newProcessingStepCalculation({ step })
     const stepWithCalculation = Step.assocCalculation(calculation)(step)
     const chainWithStep = Chain.assocProcessingStep(stepWithCalculation)(State.getChainEdit(state))
